# public/swagger.json
{
	"components": {
...
		"schemas": {
...
			"User": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"firstName": {
						"type": "string"
					},
					"lastName": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"active": {
						"type": "number",
						"format": "double"
					},
					"admin": {
						"type": "number",
						"format": "double"
					},
					"banned": {
						"type": "number",
						"format": "double"
					},
					"avatar": {
						"type": "string"
					},
					"activation_token": {
						"type": "string"
					},
					"remember_token": {
						"type": "string"
					},
					"subscribers": {
						"items": {
							"$ref": "#/components/schemas/Subscriber"
						},
						"type": "array"
					},
					"created_at": {
						"type": "string",
						"format": "date-time"
					},
					"updated_at": {
						"type": "string",
						"format": "date-time"
					},
					"message": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"firstName",
					"lastName",
					"email",
					"password",
					"active",
					"admin",
					"banned",
					"avatar",
					"activation_token",
					"remember_token",
					"subscribers",
					"created_at",
					"updated_at",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Subscriber": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"user": {
						"$ref": "#/components/schemas/User"
					},
					"is_subscribed": {
						"type": "number",
						"format": "double"
					},
					"time_limit": {
						"type": "string"
					},
					"download_limit": {
						"type": "number",
						"format": "double"
					},
					"downloaded_data": {
						"type": "number",
						"format": "double"
					},
					"special_download_limit": {
						"type": "number",
						"format": "double"
					},
					"special_downloaded_data": {
						"type": "number",
						"format": "double"
					},
					"created_at": {
						"type": "string",
						"format": "date-time"
					},
					"updated_at": {
						"type": "string",
						"format": "date-time"
					},
					"message": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"user",
					"is_subscribed",
					"time_limit",
					"download_limit",
					"downloaded_data",
					"special_download_limit",
					"special_downloaded_data",
					"created_at",
					"updated_at",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"IUserPayload": {
				"properties": {
					"firstName": {
						"type": "string"
					},
					"lastName": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					}
				},
				"required": [
					"firstName",
					"lastName",
					"email",
					"password"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Artist": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"slug": {
						"type": "string"
					},
					"image": {
						"type": "string"
					},
					"albums": {
						"items": {
							"$ref": "#/components/schemas/Album"
						},
						"type": "array"
					},
					"tracks": {
						"items": {
							"$ref": "#/components/schemas/Track"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"name",
					"slug",
					"image",
					"albums",
					"tracks"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Album": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"title": {
						"type": "string"
					},
					"slug": {
						"type": "string"
					},
					"location": {
						"type": "string"
					},
					"views": {
						"type": "number",
						"format": "double"
					},
					"downloads": {
						"type": "number",
						"format": "double"
					},
					"release_date": {
						"type": "string",
						"format": "date-time"
					},
					"featured": {
						"type": "number",
						"format": "double"
					},
					"catalog": {
						"type": "string"
					},
					"top_album": {
						"type": "number",
						"format": "double"
					},
					"vinyl_album": {
						"type": "number",
						"format": "double"
					},
					"bandcamp_album": {
						"type": "number",
						"format": "double"
					},
					"artist": {
						"$ref": "#/components/schemas/Artist"
					},
					"category": {
						"$ref": "#/components/schemas/Category"
					},
					"publisher": {
						"$ref": "#/components/schemas/Publisher"
					},
					"language": {
						"$ref": "#/components/schemas/Language"
					},
					"tracks": {
						"items": {
							"$ref": "#/components/schemas/Track"
						},
						"type": "array"
					},
					"created_at": {
						"type": "string",
						"format": "date-time"
					},
					"updated_at": {
						"type": "string",
						"format": "date-time"
					},
					"message": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"title",
					"slug",
					"location",
					"views",
					"downloads",
					"release_date",
					"featured",
					"catalog",
					"top_album",
					"vinyl_album",
					"bandcamp_album",
					"artist",
					"category",
					"publisher",
					"language",
					"tracks",
					"created_at",
					"updated_at",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Track": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"title": {
						"type": "string"
					},
					"slug": {
						"type": "string"
					},
					"location": {
						"type": "string"
					},
					"duration": {
						"type": "number",
						"format": "double"
					},
					"bpm": {
						"type": "number",
						"format": "double"
					},
					"remixed": {
						"type": "string"
					},
					"filesize": {
						"type": "number",
						"format": "double"
					},
					"downloads": {
						"type": "number",
						"format": "double"
					},
					"search": {
						"type": "string"
					},
					"artist": {
						"$ref": "#/components/schemas/Artist"
					},
					"key": {
						"$ref": "#/components/schemas/Key"
					},
					"category": {
						"$ref": "#/components/schemas/Category"
					},
					"album": {
						"$ref": "#/components/schemas/Album"
					},
					"created_at": {
						"type": "string",
						"format": "date-time"
					},
					"updated_at": {
						"type": "string",
						"format": "date-time"
					},
					"message": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"title",
					"slug",
					"location",
					"duration",
					"bpm",
					"remixed",
					"filesize",
					"downloads",
					"search",
					"artist",
					"key",
					"category",
					"album",
					"created_at",
					"updated_at",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Key": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"slug": {
						"type": "string"
					},
					"tracks": {
						"items": {
							"$ref": "#/components/schemas/Track"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"name",
					"slug",
					"tracks"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Category": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"slug": {
						"type": "string"
					},
					"albums": {
						"items": {
							"$ref": "#/components/schemas/Album"
						},
						"type": "array"
					},
					"tracks": {
						"items": {
							"$ref": "#/components/schemas/Track"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"name",
					"slug",
					"albums",
					"tracks"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Publisher": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"slug": {
						"type": "string"
					},
					"albums": {
						"items": {
							"$ref": "#/components/schemas/Album"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"name",
					"slug",
					"albums"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Language": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"slug": {
						"type": "string"
					},
					"albums": {
						"items": {
							"$ref": "#/components/schemas/Album"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"name",
					"slug",
					"albums"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"IUploadBulkPayload": {
				"properties": {
					"uploadpath": {
						"type": "string"
					}
				},
				"required": [
					"uploadpath"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ITrackPayload": {
				"properties": {
					"title": {
						"type": "string"
					},
					"slug": {
						"type": "string"
					}
				},
				"required": [
					"title",
					"slug"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "express-typescript",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
...
		"/api/v1/users/user-info": {
			"get": {
				"operationId": "GetUserInfo",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/User"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"User"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "token",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/v1/users": {
			"get": {
				"operationId": "GetUsers",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/User"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"User"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": []
			},
			"post": {
				"operationId": "CreateUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/User"
								}
							}
						}
					}
				},
				"tags": [
					"User"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IUserPayload"
							}
						}
					}
				}
			}
		},
		"/api/v1/users/{id}": {
			"get": {
				"operationId": "GetUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/User"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"User"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/User"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"User"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IUserPayload"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/User"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"User"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/api/v1/tracks": {
			"get": {
				"operationId": "GetTracks",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Track"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Track"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "pickType",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "skip",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "keyword",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "publisherSlug",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "artistSlug",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "title",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "bpmlow",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "bpmhigh",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "key",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "genre",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "label",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "artist",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"post": {
				"operationId": "CreateTrack",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Track"
								}
							}
						}
					}
				},
				"tags": [
					"Track"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ITrackPayload"
							}
						}
					}
				}
			}
		},
		"/api/v1/tracks/search": {
			"get": {
				"operationId": "GetSearchTracks",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Track"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Track"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "keyword",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "skip",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/v1/tracks/genre-tracks/{slug}": {
			"get": {
				"operationId": "GetGenreTracks",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Track"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Track"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "slug",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "skip",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/v1/tracks/uploadBulk": {
			"post": {
				"operationId": "UploadBulk",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Track"
								}
							}
						}
					}
				},
				"tags": [
					"Track"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IUploadBulkPayload"
							}
						}
					}
				}
			}
		},
		"/api/v1/tracks/{slug}": {
			"get": {
				"operationId": "GetTrack",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/Track"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"Track"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "slug",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/v1/tracks/{id}": {
			"put": {
				"operationId": "UpdateTrack",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/Track"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"Track"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ITrackPayload"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteTrack",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/Track"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"Track"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}